- refs #81965: Download map is now rendered by the class.
